+++ neurons/validator.py
@@ -17,16 +17,25 @@
 
 
 import time
+import random
+import io
+import json
+import base64
+import torch
+from typing import List
 
 # Bittensor
 import bittensor as bt
 
-# Bittensor Validator Template:
-import template
-from template.validator import forward
-
 # import base validator class which takes care of most of the boilerplate
-from template.base.validator import BaseValidatorNeuron
+from base.validator import BaseValidatorNeuron
+
+# Import forward dependencies.
+from base.protocol import Trace
+from utils.cairo_generator import generate_random_cairo_trace
+from utils.uids import get_random_uids
+from starkware.cairo.lang.cairo_constants import DEFAULT_PRIME
+from starkware.cairo.lang.compiler.cairo_compile import compile_cairo
 
 
 class Validator(BaseValidatorNeuron):
@@ -45,15 +54,56 @@ class Validator(BaseValidatorNeuron):
         self.load_state()
 
     async def forward(self):
-        """
-        Validator forward pass. Consists of:
-        - Generating the query
-        - Querying the miners
-        - Getting the responses
-        - Rewarding the miners
-        - Updating the scores
-        """
-        return await forward(self)
+        miner_uids = get_random_uids(self, k=min(self.config.neuron.sample_size, self.metagraph.n.item()))
+        trace, proof_bytes = generate_random_cairo_trace()
+
+        # For completeness, validators also need to generate proofs to ensure that miners are generating
+        # proofs of the given trace, rather than just creating any random garbage proof that will pass
+        # verification.
+
+        responses = await self.dendrite(
+            axons=[self.metagraph.axons[uid] for uid in miner_uids],
+            synapse=trace, # send in the execution trace
+            deserialize=False, # bogus responses shouldn't kill the validation flow
+            timeout=10,
+        )
+
+        def try_deserialize(item: Trace):
+            try:
+                return item.deserialize()
+            except:
+                return bytes()
+
+        responses = [try_deserialize(item) for item in responses]
+
+        # Log the results for monitoring purposes.
+        bt.logging.info(f"Received responses: {responses}")
+
+        # Adjust the scores based on responses from miners.
+        rewards = get_rewards(self, proof_bytes, responses)
+
+        bt.logging.info(f"Scored responses: {rewards}")
+        # Update the scores based on the rewards. You may want to define your own update_scores function for custom behavior.
+        self.update_scores(rewards, miner_uids)
+
+    # it's sufficient for us to check exact matches between proof bytes and pub inputs bytes.
+    # verifying the proof would be redundant at this stage, but a later update would likely make
+    # it more sensible to opt for proof verification instead of byte matching
+    def reward(proof_bytes: bytes, response_proof: bytes) -> float:
+        if proof_bytes != response_proof:
+            return 0.0
+
+        return 1.0
+
+
+    def get_rewards(
+        self,
+        proof_bytes: bytes,
+        responses: List[bytes],
+    ) -> torch.FloatTensor:
+        return torch.FloatTensor(
+            [reward(proof_bytes, response_proof) for response_proof in responses]
+        ).to(self.device)
 
 
 # The main function parses the configuration and runs the validator.
