--- template/validator/__init__.py
@@ -1,2 +1,0 @@
-from .forward import forward
-from .reward import reward
