+++ utils/__init__.py
@@ -1,3 +1,5 @@
+from . import cairo_generator
 from . import config
 from . import misc
+from . import rust
 from . import uids
