+++ utils/cairo_generator.py
@@ -16,6 +16,9 @@
 # DEALINGS IN THE SOFTWARE.
 
 import random
+import bittensor as bt
+from utils.rust import make_proof
+from utils.rust import make_trace_and_pub_inputs
 
 def generate_variable(id=None):
     return f"var_{random.randint(0, 999)}" if id is None else f"var_{id}"
@@ -60,3 +63,30 @@ def generate_cairo_program(minimum: int, maximum: int):
     program = generate_main_function(minimum, maximum)
     return program
 
+def generate_random_cairo_trace(minimum: int=10000, maximum: int=50000):
+    # Generate a random cairo program.
+    program = generate_cairo_program(minimum, maximum)
+    bt.logging.debug("Random cairo program generated.")
+
+    # Compile the program.
+    assembled = compile_cairo(code=program, prime=DEFAULT_PRIME, add_start=True)
+    bt.logging.debug("Random cairo program assembled.")
+
+    # Parse to JSON for the Rust binary to interpret.
+    program = assembled.Schema().dump(assembled)
+    del program["compiler_version"]
+    del program["main_scope"]
+    program = json.dumps(program)
+
+    # Generate trace and public inputs.
+    main_trace, pub_inputs = make_trace_and_pub_inputs(program)
+    bt.logging.debug("Trace and inputs created for random cairo program.")
+
+    proof_bytes = make_proof(main_trace, pub_inputs)
+    bt.logging.debug("Proof created for random cairo program.")
+
+    main_trace = base64.b64encode(main_trace)
+    pub_inputs = base64.b64encode(pub_inputs)
+
+    # Combine all together into a Trace for the miner.
+    return Trace(main_trace=main_trace, pub_inputs=pub_inputs), proof_bytes
